gsheets_plaid/web_server/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
plaid-python
python-dotenv
flask
google-cloud-firestore
google-cloud-secret-manager
gsheets-plaid
